refactor: isolate the params type
