+++ packages/next/src/client/components/params.ts
@@ -1,0 +1,1 @@
+export type Params = Record<string, string | string[] | undefined>
