+++ packages/next/src/server/route-matches/route-match.ts
@@ -1,3 +1,4 @@
+import type { Params } from '../../client/components/params'
 import type { RouteDefinition } from '../route-definitions/route-definition'
 
 /**
@@ -12,5 +13,5 @@ export interface RouteMatch<D extends RouteDefinition = RouteDefinition> {
    * the incoming request pathname. If a route match is returned without any
    * params, it should be considered a static route.
    */
-  readonly params: Record<string, string | string[]> | undefined
+  readonly params: Params | undefined
 }
