+++ packages/next/src/server/web/types.ts
@@ -27,7 +27,7 @@ export interface RequestData {
   }
   page?: {
     name?: string
-    params?: { [key: string]: string | string[] }
+    params?: { [key: string]: string | string[] | undefined }
   }
   url: string
   body?: ReadableStream<Uint8Array>
