+++ packages/next/src/shared/lib/router/utils/middleware-route-matcher.ts
@@ -1,6 +1,6 @@
 import type { BaseNextRequest } from '../../../../server/base-http'
 import type { MiddlewareMatcher } from '../../../../build/analysis/get-page-static-info'
-import type { Params } from './route-matcher'
+import type { Params } from '../../../../client/components/params'
 import { matchHas } from './prepare-destination'
 
 export interface MiddlewareRouteMatch {
