+++ packages/next/src/shared/lib/router/utils/prepare-destination.ts
@@ -1,7 +1,6 @@
 import type { IncomingMessage } from 'http'
 import type { Key } from 'next/dist/compiled/path-to-regexp'
 import type { NextParsedUrlQuery } from '../../../../server/request-meta'
-import type { Params } from './route-matcher'
 import type { RouteHas } from '../../../../lib/load-custom-routes'
 import type { BaseNextRequest } from '../../../../server/base-http'
 
@@ -14,6 +13,7 @@ import {
 } from '../../../../server/lib/interception-routes'
 import { NEXT_RSC_UNION_QUERY } from '../../../../client/components/app-router-headers'
 import { getCookieParser } from '../../../../server/api-utils/get-cookie-parser'
+import type { Params } from '../../../../client/components/params'
 
 /**
  * Ensure only a-zA-Z are used for param names for proper interpolating
@@ -55,7 +55,7 @@ export function matchHas(
   const params: Params = {}
 
   const hasMatch = (hasItem: RouteHas) => {
-    let value: undefined | string
+    let value
     let key = hasItem.key
 
     switch (hasItem.type) {
